refactor(tui): move rendering out of app.rs into ui.rs
CLAUDE.md documents the split as "app.rs holds mutable App state;
ui.rs renders", but the render functions (render_main_page,
render_qrcode_page, render_table_box, render_alert and the top-level
render dispatcher) lived on App while ui.rs only handled the terminal
lifecycle.

Move them to ui.rs as free functions taking &mut App, purely
mechanically (self -> app). The QR-code render cache keeps its exact
behavior: rendering still populates app.qrcode_cache keyed by the
selected index, and tick()/reset() in app.rs keep invalidating it.

App keeps only state and state-derived logic: title and qrcode_cache
become pub(crate) so ui.rs can read/fill them, and progress() becomes
pub(crate) but stays in app.rs since it is a state computation, not
rendering.
